+++ src/lib_mockup/RPC_client.ml
@@ -43,6 +43,9 @@ class mockup_ctxt (base_dir : string) (mem_only : bool)
     method generic_json_call meth ?body uri =
       local_ctxt#generic_json_call meth ?body uri
 
+    method generic_media_type_call meth ?body uri =
+      local_ctxt#generic_media_type_call meth ?body uri
+
     method call_service
         : 'm 'p 'q 'i 'o.
           (([< Resto.meth] as 'm), unit, 'p, 'q, 'i, 'o) RPC_service.t ->