+++ src/lib_proxy/RPC_client.ml
@@ -156,13 +156,47 @@ class http_local_ctxt (printer : Tezos_client_base.Client_context.printer)
         in
         if method_is_writer meth then delegate ()
         else
-          let* y = local_ctxt#generic_json_call meth ?body uri in
-          match y with
+          let* answer = local_ctxt#generic_json_call meth ?body uri in
+          match answer with
           | Ok (`Not_found _) | Error [Tezos_rpc.RPC_context.Not_found _] ->
               delegate ()
           | Ok x ->
               let* () =
                 Events.(emit done_json_call_locally) (meth_string, uri_string)
+              in
+              return_ok x
+          | Error _ as err -> Lwt.return err
+
+    method generic_media_type_call
+        : Service.meth ->
+          ?body:Data_encoding.json ->
+          Uri.t ->
+          RPC_context.generic_call_result Tezos_error_monad.Error_monad.tzresult
+          Lwt.t =
+      let open Lwt_syntax in
+      fun meth ?body uri ->
+        let meth_string = RPC_service.string_of_meth meth in
+        let uri_string = Uri.to_string uri in
+        let delegate () =
+          let* () =
+            Events.(emit delegate_media_type_call_to_http)
+              (meth_string, uri_string)
+          in
+          http_ctxt#generic_media_type_call meth ?body uri
+        in
+        if method_is_writer meth then delegate ()
+        else
+          let* answer = local_ctxt#generic_media_type_call meth ?body uri in
+          match answer with
+          | Ok (`Json (`Not_found _))
+          | Ok (`Binary (`Not_found _))
+          | Ok (`Other (_, `Not_found _))
+          | Error [Tezos_rpc.RPC_context.Not_found _] ->
+              delegate ()
+          | Ok x ->
+              let* () =
+                Events.(emit done_media_type_call_locally)
+                  (meth_string, uri_string)
               in
               return_ok x
           | Error _ as err -> Lwt.return err