+++ src/proto_alpha/lib_delegate/test/mockup_simulator/faked_client_context.ml
@@ -58,6 +58,9 @@ class faked_ctxt (hooks : Faked_services.hooks) (chain_id : Chain_id.t) :
     method generic_json_call meth ?body uri =
       local_ctxt#generic_json_call meth ?body uri
 
+    method generic_media_type_call meth ?body uri =
+      local_ctxt#generic_media_type_call meth ?body uri
+
     method call_service
         : 'm 'p 'q 'i 'o.
           (([< Resto.meth] as 'm), unit, 'p, 'q, 'i, 'o) RPC_service.t ->